refactor: extract hook lifecycle from lib.rs into hooks.rs
