+++ src/scanner.rs
@@ -178,41 +178,7 @@ pub fn run_scan(
     let mut detectors = initialize_detectors().map_err(|err| err.to_string())?;
 
     if let Some(cfg) = config {
-        if let Some(ref rules) = cfg.rules {
-            for rule in rules {
-                match Detector::new(
-                    &rule.name,
-                    &rule.pattern,
-                    &rule.finding_type,
-                    &rule.severity,
-                    &[],
-                    &[],
-                    None,
-                ) {
-                    Ok(det) => detectors.push(det),
-                    Err(e) => eprintln!("Warning: custom rule '{}' skipped: {}", rule.name, e),
-                }
-            }
-        }
-
-        if let Some(ref overrides) = cfg.overrides {
-            detectors.retain(|det| {
-                if let Some(override_config) = overrides.get(&det.name) {
-                    if override_config.enabled == Some(false) {
-                        return false;
-                    }
-                }
-                true
-            });
-
-            for det in &mut detectors {
-                if let Some(override_config) = overrides.get(&det.name) {
-                    if let Some(ref sev) = override_config.severity {
-                        det.severity.clone_from(sev);
-                    }
-                }
-            }
-        }
+        cfg.apply_to(&mut detectors);
     }
     let (multiline_detectors, line_detectors): (Vec<_>, Vec<_>) = detectors
         .iter()
@@ -309,7 +275,7 @@ pub fn run_scan(
     }
     let unique_paths: Vec<_> = unique_paths.into_iter().collect();
 
-    let exclude_patterns: Vec<Pattern> = args
+    let mut exclude_patterns: Vec<Pattern> = args
         .exclude
         .as_ref()
         .map(|exclude_str| {
@@ -325,21 +291,13 @@ pub fn run_scan(
         .transpose()?
         .unwrap_or_default();
 
-    let exclude_patterns = if let Some(cfg) = config {
-        if let Some(ref cfg_excludes) = cfg.exclude {
-            let mut patterns = exclude_patterns;
-            for pattern_str in cfg_excludes {
-                patterns.push(Pattern::new(pattern_str).map_err(|err| {
-                    format!("Invalid config exclude pattern '{}': {}", pattern_str, err)
-                })?);
-            }
-            patterns
-        } else {
-            exclude_patterns
+    if let Some(excludes) = config.and_then(|cfg| cfg.exclude.as_ref()) {
+        for pattern_str in excludes {
+            exclude_patterns.push(Pattern::new(pattern_str).map_err(|err| {
+                format!("Invalid config exclude pattern '{}': {}", pattern_str, err)
+            })?);
         }
-    } else {
-        exclude_patterns
-    };
+    }
 
     let results: Vec<(Vec<Finding>, usize, usize, Option<String>)> = unique_paths
         .into_par_iter()
